ab_contract_file

Enum ContractFileParseError

pub enum ContractFileParseError {

    FileTooLarge {
        file_size: usize,
    },
    NoHeader,
    WrongMagicBytes,
    MetadataOutOfRange {
        offset: u32,
        size: u16,
        file_size: u32,
    },
    MetadataDecoding,
    FileTooSmall {
        num_methods: u16,
        read_only_section_size: u32,
        file_size: u32,
    },
    MethodUnaligned {
        file_offset: u32,
        memory_address: u32,
    },
    MethodOutOfRange {
        offset: u32,
        code_section_offset: u32,
        file_size: u32,
    },
    HostCallFnUnaligned {
        file_offset: u32,
        memory_address: u32,
    },
    HostCallFnOutOfRange {
        offset: u32,
        code_section_offset: u32,
        file_size: u32,
    },
    InvalidHostCallFnPattern {
        instruction: ContractInstruction,
    },
    InvalidReadOnlySizes {
        file_size: u32,
        memory_size: u32,
    },
    InsufficientHeaderMethods {
        header_num_methods: u16,
        metadata_method_index: u16,
    },
    MetadataNumMethodsMismatch {
        header_num_methods: u16,
        metadata_num_methods: u16,
    },
    InvalidInstruction {
        instruction: u32,
    },
    CodeEmpty,
    UnexpectedTrailingCodeBytes {
        num_bytes: usize,
    },
    LastInstructionMustBeJump {
        instruction: ContractInstruction,
    },

}

Error for ContractFile::parse()

Variants

FileTooLarge

The file is too large, must fit into u32

Fields

file_size: usize

Actual file size

NoHeader

The file does not have a header (not enough bytes)

WrongMagicBytes

The magic bytes in the header are incorrect

MetadataOutOfRange

The metadata section is out of bounds of the file

Fields

offset: u32

Offset of the metadata section in bytes relative to the start of the file

size: u16

Size of the metadata section in bytes

file_size: u32

Size of the file in bytes

MetadataDecoding

Failed to decode metadata item

FileTooSmall

The file is too small

Fields

num_methods: u16

Number of methods in the contract

read_only_section_size: u32

Size of the read-only section in bytes as stored in the file

file_size: u32

Size of the file in bytes

MethodUnaligned

Method is unaligned

Fields

file_offset: u32

Offset of the method in bytes relative to the start of the file

memory_address: u32

Address of the method in bytes relative to the beginning of the initialized memory

MethodOutOfRange

Method offset is out of bounds of the file

Fields

offset: u32

Offset of the method in bytes relative to the start of the file

code_section_offset: u32

Offset of the code section in bytes relative to the start of the file

file_size: u32

Size of the file in bytes

HostCallFnUnaligned

Host call function is unaligned

Fields

file_offset: u32

Offset of the method in bytes relative to the start of the file

memory_address: u32

Address of the method in bytes relative to the beginning of the initialized memory

HostCallFnOutOfRange

The host call function offset is out of bounds of the file

Fields

offset: u32

Offset of the host call function in bytes relative to the start of the file

code_section_offset: u32

Offset of the code section in bytes relative to the start of the file

file_size: u32

Size of the file in bytes

InvalidHostCallFnPattern

Host call function doesn’t have jal tailcall instruction

Fields

instruction: ContractInstruction

Instruction of the host call function

InvalidReadOnlySizes

The read-only section file size is larger than the memory size

Fields

file_size: u32

Size of the read-only section in bytes as stored in the file

memory_size: u32

Size of the read-only section in bytes as will be written to memory during execution

InsufficientHeaderMethods

There are not enough methods in the header to match the number of methods in the actual metadata

Fields

header_num_methods: u16

Number of methods in the header

metadata_method_index: u16

Index of the method in the actual metadata that is missing from the header

MetadataNumMethodsMismatch

The number of methods in the header does not match the number of methods in the actual metadata

Fields

header_num_methods: u16

Number of methods in the header

metadata_num_methods: u16

Number of methods in the actual metadata

InvalidInstruction

Invalid instruction encountered while parsing the code section

Fields

instruction: u32

Instruction

CodeEmpty

The code section is empty

UnexpectedTrailingCodeBytes

Unexpected trailing code bytes encountered while parsing the code section

Fields

num_bytes: usize

Number of trailing bytes encountered

LastInstructionMustBeJump

The last instruction in the code section must be a jump instruction

Fields

instruction: ContractInstruction

Instruction that is expected to be a jump instruction