ab_contract_file

Struct ContractFile

pub struct ContractFile<'a> { /* private fields */ }

A container for a parsed contract file

Implementations

impl<'a> ContractFile<'a>

pub fn parse<CM>( file_bytes: &'a [u8], contract_method: CM, ) -> Result<Self, ContractFileParseError>
where CM: FnMut(ContractFileMethod<'a>) -> Result<(), ContractFileParseError>,

Parse file bytes and verify that internal invariants are valid.

contract_method argument is an optional callback called for each method in the contract file with its method address in the contract memory, metadata item, and corresponding metadata bytes. This can be used to collect available methods during parsing and avoid extra iteration later using Self::iterate_methods() to compute MethodFingerprint, etc.

pub unsafe fn parse_unchecked(file_bytes: &'a [u8]) -> Self

Similar to ContractFile::parse() but does not verify internal invariants and assumes the input is valid.

This method is more efficient and does no checks that ContractFile::parse() does.

Safety

Must be a valid input, for example, previously verified using ContractFile::parse().

pub fn header(&self) -> ContractFileHeader

Get file header

pub fn metadata_bytes(&self) -> &[u8]

Metadata stored in the file

pub fn contract_memory_size(&self) -> u32

Memory allocation required for the contract

pub fn initialize_contract_memory( &self, contract_memory: &mut [MaybeUninit<u8>], ) -> bool

Initialize contract memory with file contents.

Use Self::contract_memory_size() to identify the exact necessary amount of memory.

pub fn get_code(&self) -> &[u8]

Get the complete code section with instructions

pub fn iterate_methods( &self, ) -> impl ExactSizeIterator<Item = ContractFileMethod<'_>> + TrustedLen

Iterate over all methods in the contract